FIXME: cast away size_t -> x warnings

+++ base/applications/mscutils/eventvwr/eventvwr.c
@@ -1186,10 +1186,10 @@ FormatByteSize(LONGLONG cbSize, LPWSTR pwszResult, UINT cchResultMax)
     pwszEnd = pwszResult + cchWritten;
     cchRemaining = cchResultMax - cchWritten;
     StringCchCopyExW(pwszEnd, cchRemaining, L" ", &pwszEnd, &cchRemaining, 0);
-    cchWritten = LoadStringW(hInst, IDS_BYTES_FORMAT, pwszEnd, cchRemaining);
+    cchWritten = LoadStringW(hInst, IDS_BYTES_FORMAT, pwszEnd, (ULONG)cchRemaining);
     cchRemaining -= cchWritten;
 
-    return cchResultMax - cchRemaining;
+    return cchResultMax - (UINT)cchRemaining;
 }
 
 LPWSTR
@@ -1214,7 +1214,7 @@ FormatFileSizeWithBytes(const PULARGE_INTEGER lpQwSize, LPWSTR pwszResult, UINT 
     StringCchCopyExW(pwszEnd, cchRemaining, L" (", &pwszEnd, &cchRemaining, 0);
 
     /* Write formated bytes count */
-    cchWritten = FormatByteSize(lpQwSize->QuadPart, pwszEnd, cchRemaining);
+    cchWritten = FormatByteSize(lpQwSize->QuadPart, pwszEnd, (UINT)cchRemaining);
     pwszEnd += cchWritten;
     cchRemaining -= cchWritten;
 
@@ -1237,7 +1237,7 @@ GetFileTimeString(LPFILETIME lpFileTime, LPWSTR pwszResult, UINT cchResult)
     if (!FileTimeToLocalFileTime(lpFileTime, &ft) || !FileTimeToSystemTime(&ft, &st))
         return FALSE;
 
-    cchWritten = GetDateFormatW(LOCALE_USER_DEFAULT, DATE_LONGDATE, &st, NULL, pwszEnd, cchRemaining);
+    cchWritten = GetDateFormatW(LOCALE_USER_DEFAULT, DATE_LONGDATE, &st, NULL, pwszEnd, (ULONG)cchRemaining);
     if (cchWritten)
         --cchWritten; // GetDateFormatW returns count with terminating zero
     // else
@@ -1248,7 +1248,7 @@ GetFileTimeString(LPFILETIME lpFileTime, LPWSTR pwszResult, UINT cchResult)
 
     StringCchCopyExW(pwszEnd, cchRemaining, L", ", &pwszEnd, &cchRemaining, 0);
 
-    cchWritten = GetTimeFormatW(LOCALE_USER_DEFAULT, 0, &st, NULL, pwszEnd, cchRemaining);
+    cchWritten = GetTimeFormatW(LOCALE_USER_DEFAULT, 0, &st, NULL, pwszEnd, (ULONG)cchRemaining);
     if (cchWritten)
         --cchWritten; // GetTimeFormatW returns count with terminating zero
     // else
